Modify code so it is a bit cleaner and structure files into folder structure
