+++ pages/react_section/projects/robofriends/src/components/Card.js
