+++ pages/react_section/projects/robofriends/src/components/SearchBox.js
