+++ pages/react_section/projects/robofriends/src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import reportWebVitals from './reportWebVitals';
+import './index.css';
+import App from './containers/App';
 import 'tachyons';
-import App from './App';
 
 
 ReactDOM.render(
